Adjust directory composition.
