+++ supervised/softmax-mnist.py
